django3_todowoo_project/requirements/base.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
dataclasses==0.8
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==3.4.0
Markdown==3.3.3
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2020.5
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.0
